requirements.txt
Flask
pymongo
argon2-cffi
pyotp
flask-limiter
flask-talisman
Flask-Mail
tzdata
flask-wtf